+++ hydra/atmosphere.cpp
@@ -301,7 +301,7 @@ void Atmosphere::updateSkyCubemap(Cubemap &cubemap) const {
     auto colors = getSkyboxColors();
 
     if (skyboxCacheValid) {
-        double maxComponentDelta = 0.0;
+        float maxComponentDelta = 0.0;
         for (size_t face = 0; face < colors.size(); ++face) {
             const Color &newColor = colors[face];
             const Color &oldColor = lastSkyboxColors[face];
